apps/api-backend/requirements.txt
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
sqlalchemy[asyncio]>=2.0.25,<3.0
asyncpg>=0.29.0,<1.0
alembic>=1.13.0,<2.0
pydantic>=2.6.0,<3.0
pydantic-settings>=2.1.0,<3.0
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0,<2.0
httpx>=0.26.0,<1.0
pytest>=8.0.0,<9.0
pytest-asyncio>=0.23.0,<1.0
openai>=1.12.0,<2.0
langchain>=0.1.0,<1.0
azure-cognitiveservices-speech>=1.35.0,<2.0
py-fsrs>=1.0.0,<2.0
tenacity>=8.2.0,<9.0
sse-starlette>=1.8.0,<3.0
